Moving assert never to test_utils
