+++ lib/agent0/agent0/hyperdrive/crash_report/known_error_checks.py
@@ -3,7 +3,7 @@ from ethpy.base.errors import ContractCallException
 from web3.exceptions import ContractCustomError, ContractLogicError
 
 from agent0.hyperdrive.state import HyperdriveActionType, TradeResult
-from agent0.utils import assert_never
+from agent0.test_utils import assert_never
 
 
 def check_for_invalid_balance(trade_result: TradeResult) -> tuple[bool, TradeResult]:
