+++ lib/agent0/agent0/test_utils/__init__.py
@@ -1,2 +1,3 @@
 """Test fixtures for agent0"""
+from .assert_never import assert_never
 from .cycle_trade_policy import CycleTradesPolicy, cycle_trade_policy
